refactor: split `dpp::thread` from `channel.h` into `thread.h`
fixes `dpp::message`'s copy constructor calling `std::vector<dpp::channel>`'s copy constructor with incomplete `dpp::channel`
