+++ include/dpp/channel.h
@@ -26,7 +26,6 @@
 #include <dpp/managed.h>
 #include <dpp/utility.h>
 #include <dpp/voicestate.h>
-#include <dpp/message.h>
 #include <dpp/json_fwd.h>
 #include <dpp/permissions.h>
 #include <dpp/json_interface.h>
@@ -243,38 +242,6 @@ struct DPP_EXPORT permission_overwrite {
 	permission_overwrite(snowflake id, uint64_t allow, uint64_t deny, overwrite_type type);
 };
 
-
-/**
- * @brief metadata for threads
- */
-struct DPP_EXPORT thread_metadata {
-	/**
-	 * @brief Timestamp when the thread's archive status was last changed, used for calculating recent activity.
-	 */
-	time_t archive_timestamp;
-
-	/**
-	 * @brief The duration in minutes to automatically archive the thread after recent activity (60, 1440, 4320, 10080).
-	 */
-	uint16_t auto_archive_duration;
-
-	/**
-	 * @brief Whether a thread is archived
-	 */
-	bool archived;
-
-	/**
-	 * @brief Whether a thread is locked. When a thread is locked,
-	 * only users with `MANAGE_THREADS` can un-archive it.
-	 */
-	bool locked;
-
-	/**
-	 * @brief Whether non-moderators can add other non-moderators. Only for private threads.
-	 */
-	bool invitable;
-};
-
 /**
  * @brief Auto archive duration of threads which will stop showing in the channel list after the specified period of inactivity.
  * Defined as an enum to fit into 1 byte. Internally it'll be translated to minutes to match the API
@@ -299,42 +266,6 @@ enum auto_archive_duration_t : uint8_t {
 	 * @brief Auto archive duration of 1 week (10080 minutes).
 	 */
 	arc_1_week = 4,
-};
-
-/**
- * @brief represents membership of a user with a thread
- */
-struct DPP_EXPORT thread_member : public json_interface<thread_member> {
-protected:
-	friend struct json_interface<thread_member>;
-
-	/**
-	 * @brief Read struct values from a json object 
-	 * @param j json to read values from
-	 * @return A reference to self	
-	 */
-	thread_member& fill_from_json_impl(nlohmann::json* j);
-
-public:
-	/**
-	 * @brief ID of the thread member is part of.
-	 */
-	snowflake thread_id;
-
-	/**
-	 * @brief ID of the member.
-	 */
-	snowflake user_id;
-
-	/**
-	 * @brief The time when user last joined the thread.
-	 */
-	time_t joined;
-
-	/**
-	 * @brief Any user-thread settings, currently only used for notifications.
-	 */
-	uint32_t flags;
 };
 
 /**
@@ -402,11 +333,6 @@ public:
 };
 
 /**
- * @brief A group of thread member objects. the key is the user_id of the dpp::thread_member
- */
-typedef std::unordered_map<snowflake, thread_member> thread_member_map;
-
-/**
  * @brief A definition of a discord channel.
  * There are one of these for every channel type except threads. Threads are
  * special snowflakes. Get it? A Discord pun. Hahaha. .... I'll get my coat.
@@ -428,6 +354,8 @@ protected:
 	 * @return json JSON object
 	 */
 	virtual json to_json_impl(bool with_id = false) const;
+
+	static constexpr uint16_t CHANNEL_TYPE_MASK = 0b0000000000001111;
 
 public:
 	/**
@@ -937,111 +865,6 @@ public:
 
 };
 
-/** @brief A definition of a discord thread.
- * A thread is a superset of a channel. Not to be confused with `std::thread`!
- */
-class DPP_EXPORT thread : public channel, public json_interface<thread> {
-protected:
-	friend struct json_interface<thread>;
-
-	/** Read class values from json object
-	 * @param j A json object to read from
-	 * @return A reference to self
-	 */
-	thread& fill_from_json_impl(nlohmann::json* j);
-
-	/**
-	 * @brief Build json for this thread object
-	 *
-	 * @param with_id include the ID in the json
-	 * @return std::string JSON string
-	 */
-	json to_json_impl(bool with_id = false) const override;
-
-public:
-	using json_interface<thread>::fill_from_json;
-	using json_interface<thread>::build_json;
-	using json_interface<thread>::to_json;
-
-	/**
-	 * @brief Thread member of current user if joined to the thread.
-	 * Note this is only set by certain api calls otherwise contains default data
-	 */
-	thread_member member;
-
-	/**
-	 * @brief Thread metadata (threads)
-	 */
-	thread_metadata metadata;
-
-	/**
-	 * @brief Created message. Only filled within the cluster::thread_create_in_forum() method
-	 */
-	message msg;
-
-	/**
-	 * @brief A list of dpp::forum_tag IDs that have been applied to a thread in a forum or media channel.
-	 */
-	std::vector<snowflake> applied_tags;
-
-	/**
-	 * @brief Number of messages ever sent in the thread.
-	 * It's similar to thread::message_count on message creation, but will not decrement the number when a message is deleted
-	 */
-	uint32_t total_messages_sent;
-
-	/**
-	 * @brief Number of messages (not including the initial message or deleted messages) of the thread.
-	 * For threads created before July 1, 2022, the message count is inaccurate when it's greater than 50.
-	 */
-	uint8_t message_count;
-
-	/**
-	 * @brief Approximate count of members in a thread (stops counting at 50)
-	 */
-	uint8_t member_count;
-
-	/**
-	 * @brief Construct a new thread object
-	 */
-	thread();
-
-	/**
-	 * @brief Returns true if the thread is within an announcement channel
-	 *
-	 * @return true if announcement thread
-	 */
-	bool is_news_thread() const;
-
-	/**
-	 * @brief Returns true if the channel is a public thread
-	 *
-	 * @return true if public thread
-	 */
-	bool is_public_thread() const;
-
-	/**
-	 * @brief Returns true if the channel is a private thread
-	 *
-	 * @return true if private thread
-	 */
-	bool is_private_thread() const;
-
-	/**
-	 * @brief Destroy the thread object
-	 */
-	virtual ~thread() = default;
-};
-
-
-/**
- * @brief Serialize a thread_metadata object to json
- *
- * @param j JSON object to serialize to
- * @param tmdata object to serialize
- */
-void to_json(nlohmann::json& j, const thread_metadata& tmdata);
-
 /**
  * @brief Serialize a permission_overwrite object to json
  *
@@ -1054,31 +877,6 @@ void to_json(nlohmann::json& j, const permission_overwrite& po);
  * @brief A group of channels
  */
 typedef std::unordered_map<snowflake, channel> channel_map;
-
-/**
- * @brief A group of threads
- */
-typedef std::unordered_map<snowflake, thread> thread_map;
-
-/**
- * @brief A thread alongside the bot's optional thread_member object tied to it
- */
-struct active_thread_info {
-	/**
-	 * @brief The thread object
-	 */
-	thread active_thread;
-
-	/**
-	 * @brief The bot as a thread member, only present if the bot is in the thread
-	 */
-	std::optional<thread_member> bot_member;
-};
-
-/**
- * @brief A map of threads alongside optionally the thread_member tied to the bot if it is in the thread. The map's key is the thread id. Returned from the cluster::threads_get_active method
- */
-using active_threads = std::map<snowflake, active_thread_info>;
 
 } // namespace dpp
