+++ include/dpp/dispatcher.h
@@ -28,6 +28,7 @@
 #include <dpp/role.h>
 #include <dpp/user.h>
 #include <dpp/channel.h>
+#include <dpp/thread.h>
 #include <dpp/guild.h>
 #include <dpp/invite.h>
 #include <dpp/emoji.h>
